Extract class ZUGFeRDTransactionModelConverter from ZUGFeRDExporter so that it can be tested separately and improve readability
